chore(e2e): adds localStorage setup and teardown step
